# stateless-validator: Use optimized MPT