--- a/Cargo.toml
+++ b/Cargo.toml
@@ -122,6 +122,8 @@ reth-ethereum-primitives = { git = "https://github.com/kevaundray/reth", rev = "
 reth-stateless = { git = "https://github.com/kevaundray/reth", rev = "e0a902c98d586f2a306245d5942712865bce6294" }
 reth-rpc-api = { git = "https://github.com/kevaundray/reth", rev = "e0a902c98d586f2a306245d5942712865bce6294" }
 reth-primitives-traits = { git = "https://github.com/kevaundray/reth", rev = "e0a902c98d586f2a306245d5942712865bce6294" }
+reth-errors = { git = "https://github.com/kevaundray/reth", rev = "e0a902c98d586f2a306245d5942712865bce6294" }
+reth-trie-common = { git = "https://github.com/kevaundray/reth", rev = "e0a902c98d586f2a306245d5942712865bce6294" }
 
 
 # alloy
@@ -131,10 +133,15 @@ alloy-consensus = { version = "1.0.18", default-features = false }
 alloy-primitives = { version = "1.2.0", default-features = false }
 alloy-hardforks = { version = "0.2.7", default-features = false }
 alloy-chains = { version = "0.2.5", default-features = false }
+alloy-rlp = { version = "0.3.10", default-features = false }
+alloy-trie = { version = "0.9.0", default-features = false }
 
 ethereum_ssz_derive = "0.9"
 ethereum_ssz = "0.9"
 
+revm-bytecode = { version = "6.0.1", default-features = false }
+
+risc0-ethereum-trie = { git = "https://github.com/risc0/risc0-ethereum.git", rev = "4a5efe13b3655258b89e7082c57743e445133dc8" }
 
 # misc
 bincode = "1.3"