--- a/ere-guests/libs/src/lib.rs
+++ b/ere-guests/libs/src/lib.rs
@@ -9,6 +9,7 @@ use serde_with::serde_as;
 
 pub mod block_ssz;
 pub mod chainconfig;
+pub mod mpt;
 
 /// Block wrapper that supports bincode serialization
 #[serde_as]