--- a/ere-guests/libs/src/mpt.rs
+++ b/ere-guests/libs/src/mpt.rs
@@ -1,3 +1,6 @@
+//! Provides an optimized sparse MPT implementation for the stateless validator guest program.
+#![allow(warnings)]
+
 // Copyright 2025 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");