--- a/ere-guests/stateless-validator/sp1/src/main.rs
+++ b/ere-guests/stateless-validator/sp1/src/main.rs
@@ -3,7 +3,6 @@
 #![no_main]
 
 extern crate alloc;
-
 use alloc::sync::Arc;
 
 use guest_libs::{chainconfig::ChainConfig, mpt::SparseState};