… Cargo.lock
… ere-guests/Cargo.toml
… ere-guests/libs/Cargo.toml
… scripts/compare_executions.py
… scripts/compare_provings.py